:alien: Add compatibility with home assistant 0.89.0
Fix #1
